refactor: 159-move discountMark function & rename file
